requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
boto3
requests
sqlalchemy
asyncmy  # or aiomysql, depending on your MySQL driver